requirements.txt
streamlit
openai
google-cloud-bigquery
pandas
plotly
scikit-learn